Move seg (seq) num generation to deduplicator
The stats struct should only be responsible for tracking the
statistics for the segments. It should *not* be on the path
for the generation of the deduplicated stream. Eventually we
might make stats tracking optional.

This change facilitates that by moving the segment seq num
generation in to the deduplicator. The stats struct now
tracks segments based on the hash issued to it by the
dedplicator, instead of generating them itself.